+++ example/lib/place.dart
@@ -3,4 +3,10 @@ class Place {
   final bool isClosed;
 
   const Place({this.name, this.isClosed = false});
+
+  @override
+  String toString() {
+    // TODO: implement toString
+    return 'Place $name (closed : $isClosed)';
+  }
 }