+++ lib/src/cluster_item.dart
@@ -1,5 +1,3 @@
-import 'dart:math';
-
 import 'package:geohash/geohash.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 
@@ -10,11 +8,4 @@ class ClusterItem<T> {
 
   ClusterItem(this.location, {this.item})
       : geohash = Geohash.encode(location.latitude, location.longitude);
-
-  String getId() {
-    return location.latitude.toString() +
-        "_" +
-        location.longitude.toString() +
-        "_${Random().nextInt(10000)}";
-  }
 }