+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## [0.0.1] - TODO: Add release date.
+## 0.0.2
 
-* TODO: Describe initial release.
+- Add `setItems` and `addItem` methods
+- Add initial zoom
+
+## 0.0.1
+
+- Initial developers preview release.